+++ blockfrost-api/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version [0.3.1.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.3.0.0...v0.3.1.0) (2022-02-17)
+
+* Additions
+  * `/blocks/${hash_or_number}/addresses` endpoint (Affected addresses in a block)
+  * `blockTime` field to `AddressTransaction`
+  * `Ord` instances for `Address`, `AssetId`, `BlockHash`, `Epoch`, `Slot`, `TxHash`
+  * `Enum`, `Real`, `Integral` for `Slot` and `Epoch`
+
 # Version [0.3.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.2.1.0...v0.3.0.0) (2022-02-07)
 
 * Changes